+++ packages/pixi-solid/src/index.ts
@@ -1,17 +1,15 @@
 export type { ContainerProps, LeafProps, PixiComponentProps } from "./component-creation";
+export { createAsyncDelay, delay } from "./delay";
 export type { PixiEventHandlerMap } from "./event-names";
 export { PIXI_EVENT_NAMES, PIXI_SOLID_EVENT_HANDLER_NAMES } from "./event-names";
 export { onResize } from "./on-resize";
-export type { PixiApplicationProps, PixiScreenDimensions } from "./pixi-application";
+export { onTick } from "./on-tick";
+export type { PixiApplicationProps } from "./pixi-application";
 export {
-  createAsyncDelay,
-  delay,
   getPixiApp,
   getTicker,
-  onTick,
-  PixiApplication,
+  PixiApplicationProvider,
   TickerProvider,
-  usePixiScreen,
 } from "./pixi-application";
 export type { PixiCanvasProps } from "./pixi-canvas";
 export { PixiCanvas } from "./pixi-canvas";
@@ -33,3 +31,5 @@ export {
   TilingSprite,
 } from "./pixi-components";
 export type { PointAxisPropName } from "./point-property-names";
+export { usePixiScreen } from "./use-pixi-screen";
+export type { PixiScreenDimensions } from "./use-pixi-screen/pixi-screen-store";